+++ packages/nexrender-database-redis/package.json
@@ -1,0 +1,12 @@
+{
+  "name": "@nexrender/database-redis",
+  "version": "1.0.0",
+  "author": "inlife",
+  "main": "src/index.js",
+  "dependencies": {
+    "redis": "^3.1.2"
+  },
+  "publishConfig": {
+    "access": "public"
+  }
+}
