+++ packages/nexrender-server/package.json
@@ -23,7 +23,10 @@
     "micro": "^9.3.3",
     "micro-cors": "^0.1.1",
     "microrouter": "^3.1.3",
-    "redis": "^3.1.2"
+    "requireg": "^0.2.2"
+  },
+  "peerDependencies": {
+    "@nexrender/database-redis": "^1.0.0"
   },
   "publishConfig": {
     "access": "public"
