+++ packages/nexrender-server/src/helpers/database.js
@@ -1,7 +1,10 @@
-const databaseType = process.env.NEXRENDER_DATABASE_TYPE || 'disk';
 
-if (databaseType === 'redis') {
-    module.exports = require('../stores/redis.js');
+const requireg = require('requireg');
+
+const databaseType = process.env.NEXRENDER_DATABASE_PROVIDER;
+
+if (databaseType !== undefined) {
+    module.exports = requireg(`@nexrender/database-${databaseType}`);
 } else {
-    module.exports = require('../stores/disk.js');
+    module.exports = require('./disk.js');
 }
